+++ firmware/resource-monitor/src/ha_discovery.py
