+++ firmware/resource-monitor/src/mqtt_topics.py
