+++ firmware/resource-monitor/src/stat_util.py
